server/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.32
python-dotenv==0.21.0
Flask-Migrate==3.1.0
flask_jwt_extended==4.6.0
flask_cors==4.0.1
flask_socketio
eventlet
flake8==4.0.1
black==22.3.0
# flask_wtf==1.0.1
